DDF-04481 Combined the PKIHandler and the WssPKIHandler.
